requirements.txt
flask==3.0.0
flask-cors==4.0.0
snowflake-connector-python==3.6.0
gunicorn==21.2.0
PyJWT==2.8.0
cryptography==41.0.7
requests==2.31.0
